change the datatype of the variable

+++ DigicalulatorAPI/Services/CalculatorService.cs
@@ -4,7 +4,7 @@ namespace DigicalculatorAPI.Services
 {
     public class CalculatorService : ICalculatorService
     {
-        public double Add(double a, double b) => a + b;
+        public double Add(decimal a, decimal b) => a + b;
 
         public double Subtract(double a, double b) => a - b;
